fixes warnings detected by -Wsign-compare

+++ hrplib/hrpModel/Body.cpp
@@ -59,7 +59,7 @@ Body::~Body()
     delete rootLink_;
 
     // delete sensors
-    for(int sensorType =0; sensorType < numSensorTypes(); ++sensorType){
+    for(unsigned int sensorType =0; sensorType < numSensorTypes(); ++sensorType){
         int n = numSensors(sensorType);
         for(int i=0; i < n; ++i){
             Sensor* s = sensor(sensorType, i);
@@ -110,14 +110,14 @@ Body::Body(const Body& org)
     // copy sensors
     std::map<Link*, int> linkToIndexMap;
 
-    for(int i=0; i < org.linkTraverse_.numLinks(); ++i){
+    for(unsigned int i=0; i < org.linkTraverse_.numLinks(); ++i){
         Link* lnk = org.linkTraverse_[i];
         linkToIndexMap[lnk] = i;
     }
 
     int n = org.numSensorTypes();
     for(int sensorType = 0; sensorType < n; ++sensorType){
-        for(int i=0; i < org.numSensors(sensorType); ++i){
+        for(unsigned int i=0; i < org.numSensors(sensorType); ++i){
             Sensor* orgSensor = org.sensor(sensorType, i);
 	    if (orgSensor){
 	        int linkIndex = linkToIndexMap[orgSensor->link];
@@ -398,7 +398,7 @@ void Body::calcMassMatrix(dmatrix& out_M)
     }
 
     // subtract the constant term
-    for(size_t i = 0; i < out_M.cols(); ++i){
+    for(int i = 0; i < out_M.cols(); ++i){
         out_M.col(i) -= b1;
     }
 
@@ -615,8 +615,8 @@ void Body::addSensor(Sensor* sensor, int sensorType, int id ){
 
 void Body::clearSensorValues()
 {
-    for(int i=0; i < numSensorTypes(); ++i){
-        for(int j=0; j < numSensors(i); ++j){
+    for(unsigned int i=0; i < numSensorTypes(); ++i){
+        for(unsigned int j=0; j < numSensors(i); ++j){
             if(sensor(i,j))
                 sensor(i, j)->clear();
         }
@@ -817,7 +817,7 @@ bool CustomizedJointPath::calcInverseKinematics(const Vector3& end_p, const Matr
 
         std::vector<double> qorg(numJoints());
 		
-        for(int i=0; i < numJoints(); ++i){
+        for(unsigned int i=0; i < numJoints(); ++i){
             qorg[i] = joint(i)->q;
         }
 
@@ -849,7 +849,7 @@ bool CustomizedJointPath::calcInverseKinematics(const Vector3& end_p, const Matr
                 solved = true;
             } else {
                 solved = false;
-                for(int i=0; i < numJoints(); ++i){
+                for(unsigned int i=0; i < numJoints(); ++i){
                     joint(i)->q = qorg[i];
                 }
                 calcForwardKinematics();
@@ -894,7 +894,7 @@ void Body::calcCMJacobian(Link *base, dmatrix &J)
         }
         
         // assuming there is no branch between base and root
-        for (int i=1; i<jp->numJoints(); i++){
+        for (unsigned int i=1; i<jp->numJoints(); i++){
             l = jp->joint(i);
             l->subm = l->parent->m + l->parent->subm;
             l->submwc = l->parent->m*l->parent->wc + l->parent->submwc;
@@ -909,10 +909,10 @@ void Body::calcCMJacobian(Link *base, dmatrix &J)
     // compute Jacobian
     std::vector<int> sgn(numJoints(), 1);
     if (jp) {
-        for (int i=0; i<jp->numJoints(); i++) sgn[jp->joint(i)->jointId] = -1;
+        for (unsigned int i=0; i<jp->numJoints(); i++) sgn[jp->joint(i)->jointId] = -1;
     }
     
-    for (int i=0; i<numJoints(); i++){
+    for (unsigned int i=0; i<numJoints(); i++){
         Link *j = joint(i);
         switch(j->jointType){
         case Link::ROTATIONAL_JOINT:
@@ -981,7 +981,7 @@ void Body::calcAngularMomentumJacobian(Link *base, dmatrix &H)
         }
         
         // assuming there is no branch between base and root
-        for (int i=1; i<jp->numJoints(); i++){
+        for (unsigned int i=1; i<jp->numJoints(); i++){
             l = jp->joint(i);
             l->subm = l->parent->m + l->parent->subm;
             l->submwc = l->parent->m*l->parent->wc + l->parent->submwc;
@@ -996,10 +996,10 @@ void Body::calcAngularMomentumJacobian(Link *base, dmatrix &H)
     // compute Jacobian
     std::vector<int> sgn(numJoints(), 1);
     if (jp) {
-        for (int i=0; i<jp->numJoints(); i++) sgn[jp->joint(i)->jointId] = -1;
+        for (unsigned int i=0; i<jp->numJoints(); i++) sgn[jp->joint(i)->jointId] = -1;
     }
     
-    for (int i=0; i<numJoints(); i++){
+    for (unsigned int i=0; i<numJoints(); i++){
         Link *j = joint(i);
         switch(j->jointType){
         case Link::ROTATIONAL_JOINT:
